+++ apps/minifront/src/components/ibc/ibc-in/asset-utils.test.ts
@@ -1,7 +1,11 @@
 import { describe, expect, test } from 'vitest';
 import { fromDisplayAmount, toDisplayAmount } from './asset-utils';
+import { bigNumConfig } from '@penumbra-zone/types/lo-hi';
+import { BigNumber } from 'bignumber.js';
 
-const asset = {
+BigNumber.config(bigNumConfig);
+
+const osmoMetadata = {
   denom_units: [
     { denom: 'osmo', exponent: 6 },
     { denom: 'uosmo', exponent: 0 },
@@ -12,49 +16,67 @@ const asset = {
   symbol: 'OSMO',
 };
 
+const usdyMetadata = {
+  denom_units: [
+    {
+      denom: 'ausdy',
+      exponent: 0,
+      aliases: ['attousdy'],
+    },
+    {
+      denom: 'usdy',
+      exponent: 18,
+    },
+  ],
+  base: 'ausdy',
+  display: 'usdy',
+  name: 'Ondo US Dollar Yield',
+  symbol: 'USDY',
+};
+
 describe('toDisplayAmount', () => {
   test('converts uosmo to osmo correctly', () => {
-    expect(toDisplayAmount(asset, { denom: 'uosmo', amount: '41000000' })).toEqual('41');
+    expect(toDisplayAmount(osmoMetadata, { denom: 'uosmo', amount: '41000000' })).toEqual('41');
   });
 
   test('high precision conversion from uosmo to osmo', () => {
-    expect(toDisplayAmount(asset, { denom: 'uosmo', amount: '123456789012345' })).toEqual(
+    expect(toDisplayAmount(osmoMetadata, { denom: 'uosmo', amount: '123456789012345' })).toEqual(
       '123456789.012345',
     );
   });
 
   test('coin denom not found in asset denom_units', () => {
-    expect(toDisplayAmount(asset, { denom: 'xosmo', amount: '1000000' })).toEqual('1000000');
+    expect(toDisplayAmount(osmoMetadata, { denom: 'xosmo', amount: '1000000' })).toEqual('1000000');
   });
 
   test('zero amount conversion from uosmo to osmo', () => {
-    expect(toDisplayAmount(asset, { denom: 'uosmo', amount: '0' })).toEqual('0');
+    expect(toDisplayAmount(osmoMetadata, { denom: 'uosmo', amount: '0' })).toEqual('0');
   });
 });
 
 describe('fromDisplayAmount', () => {
   test('converts osmo to uosmo correctly for a whole number', () => {
-    const result = fromDisplayAmount(asset, 'osmo', '1');
+    const result = fromDisplayAmount(osmoMetadata, 'osmo', '1');
     expect(result).toEqual({ denom: 'uosmo', amount: '1000000' });
   });
 
   test('converts osmo to uosmo correctly for a decimal number', () => {
-    const result = fromDisplayAmount(asset, 'osmo', '0.5');
+    const result = fromDisplayAmount(osmoMetadata, 'osmo', '0.5');
     expect(result).toEqual({ denom: 'uosmo', amount: '500000' });
   });
 
   test('handles large numbers', () => {
-    const result = fromDisplayAmount(asset, 'osmo', '123456');
+    const result = fromDisplayAmount(osmoMetadata, 'osmo', '123456');
     expect(result).toEqual({ denom: 'uosmo', amount: '123456000000' });
   });
 
   test('converts when display amount is zero', () => {
-    const result = fromDisplayAmount(asset, 'osmo', '0');
+    const result = fromDisplayAmount(osmoMetadata, 'osmo', '0');
     expect(result).toEqual({ denom: 'uosmo', amount: '0' });
   });
 
   test('returns input amount if display exponent is undefined', () => {
-    const result = fromDisplayAmount(asset, 'xosmo', '100');
+    const result = fromDisplayAmount(osmoMetadata, 'xosmo', '100');
     expect(result).toEqual({ denom: 'xosmo', amount: '100' });
   });
 
@@ -68,5 +90,10 @@ describe('fromDisplayAmount', () => {
     };
     const result = fromDisplayAmount(noExponentForBase, 'osmo', '100');
     expect(result).toEqual({ denom: 'uosmo', amount: '100000000' });
+  });
+
+  test('should work with very large numbers', () => {
+    const result = fromDisplayAmount(usdyMetadata, 'usdy', '112');
+    expect(result).toEqual({ denom: 'ausdy', amount: '112000000000000000000' });
   });
 });