+++ apps/minifront/src/state/send/index.ts
@@ -161,7 +161,7 @@ const assembleRequest = ({
     value: {
       amount: toBaseUnit(
         BigNumber(amount),
-        getDisplayDenomExponentFromValueView.optional()(selection?.balanceView),
+        getDisplayDenomExponentFromValueView.optional(selection?.balanceView),
       ),
       assetId: getAssetIdFromValueView(selection?.balanceView),
     },