+++ packages/getters/src/utils/getter.ts
@@ -1,36 +1,46 @@
-/**
- * Given a value of type `SourceType`, returns a (possibly nested) property of
- * that value, of type `TargetType`. If `Optional` is `true`, returns
- * undefined if the property or an ancestor is undefined; if `false`, throws
- * when the property or an ancestor is undefined.
- */
-type GetterFunction<SourceType, TargetType, Optional extends boolean> = (
-  value: SourceType | undefined,
-) => Optional extends true ? TargetType | undefined : TargetType;
-
-interface GetterMethods<SourceType, TargetType, Optional extends boolean = false> {
+export interface Getter<SourceType, TargetType> {
   /**
-   * Returns a getter that, when given a value of type `SourceType`, returns a
-   * (possibly nested) property of that value, of type `TargetType`. If the
-   * property or any of its ancestors are undefined, returned undefined.
+   * Given an input value of `SourceType`, asserts successful retrieval of a
+   * value of `TargetType`, by the naive retrieval function passed to
+   * `createGetter`.
+   *
+   * If undefined access occurs while retrieving `TargetType`, a
+   * `GetterMissingValueError` is thrown.
+   */
+  (value?: SourceType): TargetType;
+
+  /**
+   * Every getter contains an `optional` getter, which adds `undefined` to the
+   * target type, and which will not throw `GetterMissingValueError`.
    *
    * @example
    * ```ts
-   * const getMetadata = createGetter<ValueView, Metadata>(valueView =>
+   * const getMetadataFromValueView = createGetter<ValueView, Metadata>(valueView =>
    *   valueView?.valueView.case === 'knownAssetId' ? valueView.valueView.value.metadata : undefined,
    * );
    *
-   * // Note that `valueView` has no metadata, nor even a `case`.
-   * const valueView = new ValueView();
+   * // Note that this `emptyValueView` has no metadata, nor even a `case`.
+   * const emptyValueView = new ValueView();
    *
    * // Doesn't throw, even though the metadata is missing.
-   * const metadata = getMetadata.optional()(valueView);
+   * const noMetadata: Metadata | undefined = getMetadataFromValueView.optional(emptyValueView);
    * ```
    */
-  optional: () => Getter<SourceType, TargetType, true>;
+  readonly optional: Getter<SourceType, TargetType | undefined>;
 
   /**
-   * Pipes the output of this getter to another getter or getter function.
+   * Every getter contains a `required` getter, which wraps `NonNullable` around
+   * the target type, and which will throw `GetterMissingValueError` if the
+   * accessed result is `undefined`.
+   */
+  readonly required: Getter<SourceType, NonNullable<TargetType>>;
+
+  /**
+   * Call `pipe` to create a getter for the return type of another getter or
+   * selector function provided as a parameter.  Your parameter function must
+   * accept the output of this getter as its 0th input parameter.  If the return
+   * type of your pipe parameter includes `undefined`, the return type of the
+   * created getter will also include undefined (it will be an optional getter).
    *
    * @example
    * ```ts
@@ -38,19 +48,16 @@ interface GetterMethods<SourceType, TargetType, Optional extends boolean = false
    * // if any step in the pipe is undefined.
    * const assetId1 = getMetadata.pipe(getAssetId).pipe(getInner)(valueView);
    * // Gets the deeply nested `inner` property in a metadata object, or returns
-   * // undefined if any step in the pipe is undefined. Note that `.optional()`
-   * // must be called at the _beginning_ of the chain.
-   * const assetId2 = getMetadata.optional().pipe(getAssetId).pipe(getInner)(valueView);
+   * // `undefined`.
+   * const assetId2 = getMetadata.optional.pipe(getAssetId).pipe(getInner)(valueView);
    * ```
    */
-  pipe: <PipedTargetType = unknown>(
-    pipedGetter: Getter<TargetType, PipedTargetType, Optional>,
-  ) => Getter<SourceType, PipedTargetType, Optional>;
+  readonly pipe: <PipeTargetType>(
+    pipeSelector:
+      | Getter<TargetType, PipeTargetType>
+      | ((value?: TargetType | NonNullable<TargetType> | undefined) => PipeTargetType),
+  ) => Getter<
+    SourceType,
+    TargetType extends undefined ? PipeTargetType | undefined : PipeTargetType
+  >;
 }
-
-export type Getter<
-  SourceType = unknown,
-  TargetType = unknown,
-  Optional extends boolean = false,
-> = GetterFunction<SourceType, TargetType, Optional> &
-  GetterMethods<SourceType, TargetType, Optional>;