+++ packages/services/src/view-service/unbonding-tokens-by-address-index/index.ts
@@ -44,7 +44,7 @@ export const unbondingTokensByAddressIndex: Impl['unbondingTokensByAddressIndex'
       }
 
       const regexResult = assetPatterns.unbondingToken.capture(
-        getDisplayFromBalancesResponse(new BalancesResponse(balancesResponse)) ?? '',
+        getDisplayFromBalancesResponse.optional(new BalancesResponse(balancesResponse)) ?? '',
       );
       if (!regexResult) {
         throw new Error('expected delegation token identity key not present');