+++ apps/minifront/CHANGELOG.md
@@ -1,5 +1,57 @@
 # minifront
 
+## 6.22.0
+
+### Minor Changes
+
+- 74527df: modify precision for sync percentage
+
+### Patch Changes
+
+- 0b85eca: Sort the balances by priority in the swap and send pages
+- Updated dependencies [3a5c074]
+- Updated dependencies [3a5c074]
+  - @penumbra-zone/getters@17.0.0
+  - @penumbra-zone/perspective@30.0.0
+  - @penumbra-zone/types@23.0.0
+  - @penumbra-zone/ui@9.1.1
+  - @penumbra-zone/crypto-web@24.0.0
+
+## 6.21.0
+
+### Minor Changes
+
+- 74e89e0: UI:
+
+  - Add new `Progress` component
+  - Add `MenuItem` component that shares the styles with `DropdownMenu.Item`
+  - Update the `Pill` component to support `context` prop
+  - Update the `Tabs` component to support the `compact` density
+  - Allow passing custom icons to the `Button`
+  - Fix `density` tag in Storybook
+
+  Tailwind Config:
+
+  - Add support for v2 colors with v2 prefix like `bg-v2-secondary-dark`
+
+  Minifront:
+
+  - Add top navigation to the v2 minifront with sync bar and prax connection infos
+
+### Patch Changes
+
+- Updated dependencies [74e89e0]
+- Updated dependencies [e01d5f8]
+- Updated dependencies [516caf3]
+- Updated dependencies [a3bef37]
+- Updated dependencies [5100518]
+- Updated dependencies [d938456]
+  - @penumbra-zone/ui@9.1.0
+  - @penumbra-zone/types@22.0.0
+  - @penumbra-zone/client@18.1.0
+  - @penumbra-zone/crypto-web@23.0.0
+  - @penumbra-zone/perspective@29.0.0
+
 ## 6.20.0
 
 ### Minor Changes