+++ packages/services/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @penumbra-zone/router
 
+## 34.0.0
+
+### Patch Changes
+
+- Updated dependencies [3a5c074]
+- Updated dependencies [3a5c074]
+- Updated dependencies [990291f]
+  - @penumbra-zone/getters@17.0.0
+  - @penumbra-zone/wasm@28.0.0
+  - @penumbra-zone/query@31.0.0
+  - @penumbra-zone/storage@30.0.0
+  - @penumbra-zone/types@23.0.0
+  - @penumbra-zone/crypto-web@24.0.0
+
+## 33.0.0
+
+### Patch Changes
+
+- Updated dependencies [e01d5f8]
+  - @penumbra-zone/query@30.0.0
+  - @penumbra-zone/types@22.0.0
+  - @penumbra-zone/wasm@27.0.0
+  - @penumbra-zone/crypto-web@23.0.0
+  - @penumbra-zone/storage@29.0.0
+
 ## 32.0.0
 
 ### Patch Changes